Migrate verifier
